trading_system/requirements.txt
# Trading system - Strategy B/C
numpy>=1.20
pandas>=1.3
python-dotenv>=0.19
